Squashed commit of the following:
commit 2dca86e6eed9ee902a6941fb0317d30623d81cd6
Author: Fadhil Abubaker <fadhil.abubaker@gmail.com>
Date:   Wed May 24 11:23:33 2023 -0400

    Switch rect count type from unsigned to uint8_t

+++ src/include/nirtreedisk/nirtreedisk.h
@@ -114,10 +114,10 @@ public:
         continue;
       }
 
-      unsigned rect_count = *(unsigned *)(buffer + offset);
-      offset += sizeof(unsigned);
+      uint8_t rect_count = *(uint8_t *)(buffer + offset);
+      offset += sizeof(uint8_t);
 
-      if (rect_count == std::numeric_limits<unsigned>::max()) {
+      if (rect_count == std::numeric_limits<uint8_t>::max()) {
         offset += sizeof(tree_node_handle);
       } else {
         for (unsigned r = 0; r < rect_count; r++) {
